Fixing sign and unsigned mixing.

+++ numcosmo/math/ncm_abc.c
@@ -1017,7 +1017,7 @@ ncm_abc_run (NcmABC *abc, guint nparticles)
   if (!self->started)
     g_error ("ncm_abc_run: run not started, run ncm_abc_start_run() first.");
 
-  if (nparticles <= (self->cur_sample_id + 1))
+  if (nparticles <= (guint) (self->cur_sample_id + 1))
   {
     if (self->mtype > NCM_FIT_RUN_MSGS_NONE)
     {
